Add PWA manifest.json and move service worker code to correct directory
